resources/academic_research_assistant/requirements.txt
google-adk
fastapi
uvicorn[standard]
python-dotenv
pymupdf
semanticscholar
requests
pydantic
redis>=4.6.0
